requirements.txt
# Discord - install these in order
aiohttp>=3.9.1
PyNaCl>=1.4.0
discord.py[voice]==2.3.2  # Pin to exactly version 2.3.2
git+https://github.com/imayhaveborkedit/discord-ext-voice-recv.git@main

# Environment and logging
python-dotenv>=1.0.0
loguru>=0.7.2
colorama>=0.4.6

# AI Services
openai>=1.3.0
certifi>=2023.7.22

# Pipeline
pipecat>=0.1.0  # Adjust version as needed
